+++ CHANGES.md
@@ -4,6 +4,9 @@ Full changelog
 v0.9.0 (unreleased)
 -----------------
 
+* Make sure that the scatter layer artist style editor is shown when overplotting
+  a scatter plot on top of an image.
+
 * Fix compatibility of test suite with pytest 3.x. [#1116]
 
 * Updated bundled version of WCSAxes to v0.9.